+++ crispy_forms/layout.py
@@ -584,9 +584,8 @@ class Fieldset(LayoutObject):
             legend = SafeString("")
 
         template = self.get_template_name(template_pack)
-        context = context.flatten()
         context.update({"fieldset": self, "legend": legend, "fields": fields})
-        return render_to_string(template, context)
+        return render_to_string(template, context.flatten())
 
 
 class MultiField(LayoutObject):
